+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/src/Customizations/ApplicationInsightsRestClient.cs
@@ -5,7 +5,9 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+
 using Azure.Core;
+using Azure.Monitor.OpenTelemetry.Exporter.Internals;
 using Azure.Monitor.OpenTelemetry.Exporter.Models;
 
 namespace Azure.Monitor.OpenTelemetry.Exporter
@@ -90,7 +92,7 @@ namespace Azure.Monitor.OpenTelemetry.Exporter
                 content.WriteNewLine();
             }
             request.Content = RequestContent.Create(content.ToBytes());
-            TelemetryDebugWriter.Writer.WriteTelemetry(content);
+            TelemetryDebugWriter.WriteTelemetry(content);
             return message;
         }
 
@@ -108,7 +110,7 @@ namespace Azure.Monitor.OpenTelemetry.Exporter
             request.Headers.Add("Accept", "application/json");
             using var content = new NDJsonWriter();
             request.Content = RequestContent.Create(body);
-            TelemetryDebugWriter.Writer.WriteTelemetry(content);
+            TelemetryDebugWriter.WriteTelemetry(content);
             return message;
         }
     }